Leaderboard alpha + ball speed limit + quality of life

+++ Assets/_Scripts/Supabase/SupabaseController.cs
@@ -9,7 +9,7 @@ public class SupabaseController : MonoBehaviour
     public const string API_KEY = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6ImdjZGZxenZlb2J5bHlvbnd5ZHhyIiwicm9sZSI6ImFub24iLCJpYXQiOjE3NjQ3MTc2NTgsImV4cCI6MjA4MDI5MzY1OH0.Qq8F6bScU_qy241N1UXcml1THbtxvuRSQ8vuhE00dPg";
     public string userId;
     public string username;
-    public long highscore;
+    public int highscore;
 
     private void Awake()
     {
